% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{korea}
\alias{korea}
\title{Polygons of administration area of Republic of Korea.}
\format{
A sf with 3487 rows and 4 variables:
\describe{
  \item{name}{name of administration area}
  \item{hcode_7}{id codes with 7 digit number of administration area}
  \item{hcode_10}{id codes with 10 digit number of administration area}
  \item{geometry}{wgs84 base coordinated polygons}
  ...
}
}
\source{
\url{https://github.com/vuski/admdongkor}
}
\usage{
korea
}
\description{
A dataset containing the wgs84 coordinated polygons and other attributes.
}
\keyword{datasets}
