% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seoul}
\alias{seoul}
\title{Polygons of 25 administration area of Seoul, Republic of Korea.}
\format{
A sf with 25 rows and 3 variables:
\describe{
  \item{name}{id codes with 4 digit number}
  \item{value}{numbers of sub-administration area}
  \item{geometry}{wgs84 base coordinated polygons}
  ...
}
}
\source{
\url{https://github.com/vuski/admdongkor}
}
\usage{
seoul
}
\description{
A dataset containing the wgs84 coordinated polygons and other attributes.
}
\keyword{datasets}
