% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seoul_h3}
\alias{seoul_h3}
\title{H3 addresses within Seoul, Republic of Korea.}
\format{
A data.frame with 1329 rows and 2 variables:
\describe{
  \item{name}{h3 resolution 8 address}
  \item{value}{meaningless number}
  ...
}
}
\source{
\url{https://github.com/vuski/admdongkor}
}
\usage{
seoul_h3
}
\description{
A dataset containing the h3 resolution level 8 addresses and other attributes.
}
\keyword{datasets}
