% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suwon}
\alias{suwon}
\title{administration area id code of Suwon, Republic of Korea.}
\format{
A sf with 44 rows and 2 variables:
\describe{
  \item{name}{id codes with 7 digit number of administration area}
  \item{value}{district number}
  ...
}
}
\source{
\url{https://github.com/vuski/admdongkor}
}
\usage{
suwon
}
\description{
A dataset containing 7 digit number of administration area codes and other attributes.
}
\keyword{datasets}
